requirements.txt
huggingface-hub
langchain
python-dotenv
streamlit

